requirements.txt
fastapi
passlib
Pillow
pydantic
python-dotenv
SQLAlchemy
uvicorn
psycopg2-binary
transformers
pgvector-python
python-multipart
scipy